Check entry type as part of item type checking.
